Updated CMALCr reasoner.
Included debugging messages for the reasoner.

Updated the Clauses to work with the updated version of the CMALCr 
reasoner.

Updated ClauseSet and ConnectionList to include information about the 
variables that will be used upon connection on the connection clause.

Updated the Path class for usage with the updated version of the
CMALCr reasoner.